Fix some signed unsigned compiler warnings.

+++ src/actions.cpp
@@ -1940,7 +1940,7 @@ void recalculate_fog(const gamemap& map,
 		unit_map& units,
 		std::vector<team>& teams, int team) {
 
-	assert(team >= 0 && team < teams.size());
+	assert(team >= 0 && static_cast<size_t>(team) < teams.size());
 	teams[team].refog();
 
 	for(unit_map::iterator i = units.begin(); i != units.end(); ++i) {
